fixed missing MIGRATIONS that makes laravel work

+++ netflix-api/database/migrations/2025_01_16_145226_add_times_watched_to_profiles_watched_medias_table.php
@@ -7,15 +7,21 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration {
     public function up()
     {
-        Schema::table('profiles_watched_medias', function (Blueprint $table) {
-            $table->integer('times_watched')->default(0)->after('pause_spot');
-        });
+        // Check if the column already exists
+        if (!Schema::hasColumn('Profile_Watched_Media', 'times_watched')) {
+            Schema::table('Profile_Watched_Media', function (Blueprint $table) {
+                $table->integer('times_watched')->default(0)->after('pause_spot');
+            });
+        }
     }
 
     public function down()
     {
-        Schema::table('profiles_watched_medias', function (Blueprint $table) {
-            $table->dropColumn('times_watched');
-        });
+        // Check if the column exists before dropping it
+        if (Schema::hasColumn('Profile_Watched_Media', 'times_watched')) {
+            Schema::table('Profile_Watched_Media', function (Blueprint $table) {
+                $table->dropColumn('times_watched');
+            });
+        }
     }
 };
